server/requirements.txt
FlaskFlask-PyMongo
Flask
python-dotenv